requirements.txt
flask
flask-cors
google-generativeai
opencv-python
pyttsx3
